Delete residue files
